+++ CotEditor/Sources/Document Window/Accessory Views/Advanced Character Count/AdvancedCharacterCounterView.swift
@@ -85,7 +85,7 @@ struct AdvancedCharacterCounterView: View {
         .padding(.horizontal, 10)
         .padding(.vertical, 8)
         .background(.regularMaterial)
-        .clipShape(RoundedRectangle(cornerRadius: 8))
+        .clipShape(.rect(cornerRadius: 8))
         .shadow(radius: 4, y: 2)
         .onTapGesture { }  // avoid clicking through
         .contextMenu {